refactor: Simplify file tree
