+++ lib/get-commits.js
