+++ lib/get-config.js
