+++ lib/get-next-version.js
