+++ lib/get-registry.js
