+++ lib/get-version-head.js
