+++ lib/github-release.js
