+++ lib/plugin-noop.js
