+++ lib/plugins.js
