+++ lib/publish-npm.js
