+++ lib/verify-auth.js
