+++ lib/verify-pkg.js
